packages/ai_service/requirements.txt
fastapi==0.115.2
uvicorn==0.31.1
python-multipart==0.0.12
pydantic==2.9.2
Pillow==10.4.0
opencv-python-headless==4.10.0.84
numpy==2.1.3
orjson==3.10.7
ultralytics==8.3.39
easyocr==1.7.2

